Move field specific classes to a nested directory
