+++ mathesar_ui/src/systems/data-forms/form-maker/data-form-utilities/fields/ScalarField.ts
@@ -2,14 +2,14 @@ import { _ } from 'svelte-i18n';
 
 import type { RawScalarDataFormField } from '@mathesar/api/rpc/forms';
 
+import type { DataFormStructureChangeEventHandler } from '../DataFormStructureChangeEventHandler';
+
 import {
   AbstractColumnBasedField,
   type AbstractColumnBasedFieldModifiableProps,
   type AbstractColumnBasedFieldProps,
 } from './AbstractColumnBasedField';
-import type { DataFormStructureChangeEventHandler } from './DataFormStructureChangeEventHandler';
 import type { FormFields } from './FormFields';
-import type { FormSource } from './FormSource';
 
 interface ScalarFieldProps extends AbstractColumnBasedFieldProps {
   kind: RawScalarDataFormField['kind'];
@@ -48,29 +48,5 @@ export class ScalarField extends AbstractColumnBasedField {
       ...this.getBaseFieldRawJson(),
       kind: 'scalar_column',
     };
-  }
-
-  static factoryFromRawInfo(
-    props: {
-      parentTableOid: number;
-      rawField: RawScalarDataFormField;
-    },
-    formSource: FormSource,
-  ) {
-    const { rawField } = props;
-    const baseProps = super.getBasePropsFromRawDataFormField(props, formSource);
-
-    return (
-      container: FormFields,
-      changeEventHandler: DataFormStructureChangeEventHandler,
-    ) =>
-      new ScalarField(
-        container,
-        {
-          ...baseProps,
-          kind: rawField.kind,
-        },
-        changeEventHandler,
-      );
   }
 }
