+++ mathesar_ui/src/systems/data-forms/form-maker/data-form-utilities/fields/index.ts
@@ -1,0 +1,16 @@
+export type { DataFormFieldFkInputValueHolder } from './FieldValueHolder';
+export {
+  type DataFormField,
+  type ParentDataFormField,
+  type DataFormFieldFactory,
+  buildDataFormFieldFactory,
+  buildFormFieldContainerFactory,
+} from './factories';
+export type { FkField, FkFieldPropChangeEvent } from './FkField';
+export { FieldColumn } from './FieldColumn';
+export type {
+  DataFormFieldContainerFactory,
+  FormFields,
+  FormFieldContainerChangeEvent,
+} from './FormFields';
+export type { ScalarField, ScalarFieldPropChangeEvent } from './ScalarField';
