+++ mathesar_ui/src/systems/data-forms/form-maker/elements/FormFieldElementWrapper.svelte
@@ -1,9 +1,9 @@
 <script lang="ts">
-  import type { DataFormField } from '../data-form-utilities/DataFormField';
   import {
     type DataFormManager,
     EditableDataFormManager,
   } from '../data-form-utilities/DataFormManager';
+  import type { DataFormField } from '../data-form-utilities/fields';
 
   import { AddField } from './add-field';
   import FormFieldSource from './FormFieldSource.svelte';
