+++ mathesar_ui/src/systems/data-forms/form-maker/elements/FormFieldSource.svelte
@@ -2,8 +2,8 @@
   import ColumnName from '@mathesar/components/column/ColumnName.svelte';
   import TableName from '@mathesar/components/TableName.svelte';
 
-  import type { DataFormField } from '../data-form-utilities/DataFormField';
   import type { EditableDataFormManager } from '../data-form-utilities/DataFormManager';
+  import type { DataFormField } from '../data-form-utilities/fields';
 
   export let dataFormManager: EditableDataFormManager;
   export let dataFormField: DataFormField;
