+++ mathesar_ui/src/systems/data-forms/form-maker/inspector/FieldNavigation.svelte
@@ -4,9 +4,9 @@
   import { iconExpandRight } from '@mathesar/icons';
   import { Button, Icon, iconShowMore } from '@mathesar-component-library';
 
-  import type { DataFormField } from '../data-form-utilities/DataFormField';
   import type { EditableDataFormManager } from '../data-form-utilities/DataFormManager';
   import { DataFormStructure } from '../data-form-utilities/DataFormStructure';
+  import type { DataFormField } from '../data-form-utilities/fields';
 
   import FieldNavigationElement from './FieldNavigationElement.svelte';
