+++ mathesar_ui/src/systems/data-forms/form-maker/inspector/field-config/FieldAppearance.svelte
@@ -8,7 +8,7 @@
     isDefinedNonNullable,
   } from '@mathesar-component-library';
 
-  import type { DataFormField } from '../../data-form-utilities/DataFormField';
+  import type { DataFormField } from '../../data-form-utilities/fields';
 
   export let field: DataFormField;
