+++ mathesar_ui/src/systems/data-forms/form-maker/inspector/field-config/FieldConfig.svelte
@@ -12,8 +12,8 @@
     isDefinedNonNullable,
   } from '@mathesar-component-library';
 
-  import type { DataFormField } from '../../data-form-utilities/DataFormField';
   import type { EditableDataFormManager } from '../../data-form-utilities/DataFormManager';
+  import type { DataFormField } from '../../data-form-utilities/fields';
   import FkFormFieldRuleSelector from '../../elements/FkFormFieldRuleSelector.svelte';
 
   import FieldAppearance from './FieldAppearance.svelte';
